De-duplicate code in user and home page
